TODO: fix ig

+++ Lightning_detection/V2/test.py
@@ -148,8 +148,7 @@ def get_images(startTime, endTime, storage_limit, camera, counter, sequence, out
                 #! How the hell should I do this?
                 image = cv2.resize(frames[i], size)
                 image = image[:, :, ::-1]
-                image = image / 255.0
-                image = image.astype(np.float32)
+                image = image.astype("float32") / 255.0
 
                 input_tensor_shape = interpreter.get_input_details()[0].get('shape')
                 image = np.reshape(image, input_tensor_shape)
